+++ src/Components/Auth/AuthRegister.jsx
@@ -1,0 +1,77 @@
+// Auth/AuthRegister.js
+import React, { useEffect, useState } from "react";
+import { checkUser, createUser } from "./AuthService";
+import AuthForm from "./AuthForm";
+import { useNavigate } from "react-router-dom";
+
+const AuthRegister = () => {
+  const navigate = useNavigate();
+  const [newUser, setNewUser] = useState({
+    firstName: "",
+    lastName: "",
+    email: "",
+    password: ""
+  });
+  const [add, setAdd] = useState(false);
+
+  // Redirect if the user is already logged in
+  useEffect(() => {
+    if (checkUser()) {
+      alert("You are already logged in");
+      navigate("/");
+    }
+  }, [navigate]);
+
+  // Process registration when the 'add' flag is triggered
+  useEffect(() => {
+    if (newUser && add) {
+      createUser(newUser).then((userCreated) => {
+        if (userCreated) {
+          alert(
+            `${userCreated.get("firstName")}, you successfully registered!`
+          );
+          navigate("/");
+        }
+        // Reset flag after registration
+        setAdd(false);
+      });
+    }
+  }, [navigate, newUser, add]);
+
+  const onChangeHandler = (e) => {
+    e.preventDefault();
+    const { name, value: newValue } = e.target;
+    setNewUser({
+      ...newUser,
+      [name]: newValue
+    });
+  };
+
+  const onSubmitHandler = (e) => {
+    e.preventDefault();
+    setAdd(true);
+  };
+
+  return (
+    <div className="container my-5">
+      <div className="row justify-content-center">
+        <div className="col-md-6">
+          <div className="card shadow">
+            <div className="card-header text-center bg-success text-white">
+              <h3>Register</h3>
+            </div>
+            <div className="card-body">
+              <AuthForm
+                user={newUser}
+                onChange={onChangeHandler}
+                onSubmit={onSubmitHandler}
+              />
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default AuthRegister;